+++ lib/whois/record/parser/whois.nic.nu.rb
@@ -30,12 +30,25 @@ module Whois
       #
       class WhoisNicNu < Base
 
+
+        # == Values for Status
+        #
+        # @see http://www.nic.nu/about/about5.pdf
+        #
         property_supported :status do
           if content_for_scanner =~ /Record status:\s+(.*)\n/
             case $1.downcase
-              when "active" then :registered
-              else
-                Whois.bug!(ParserError, "Unknown status `#{$1}'.")
+            when "active"
+              :registered
+            # The day after the domain name expires, its status is changed to "Not Renewed",
+            # but it is still available for you to renew during another fourweek period.
+            # After this expiration reprieve, the name is put up for shortterm auction.
+            # If no one registers the name via the auction, it expires and becomes available
+            # to the general public for  registration using the standard process.
+            when "notrenewed"
+              :redemption
+            else
+              Whois.bug!(ParserError, "Unknown status `#{$1}'.")
             end
           else
             :available